style: 데이터 불러오는 위치 변경
데이터를 지도와 캘린더 모두 데이터를 전달하기 위해
상위 컴포넌트인 StackNavigation 폴더로 코드 이동
